+++ app/models/ocom_carousels_banner.rb
@@ -1,0 +1,2 @@
+class OcomCarouselsBanner < ApplicationRecord
+end