+++ db/migrate/20220725082208_create_ocom_carousels.rb
@@ -1,0 +1,9 @@
+class CreateOcomCarousels < ActiveRecord::Migration[7.0]
+  def change
+    create_table :ocom_carousels, id: :uuid do |t|
+      t.string :title
+      t.string :description
+      t.timestamps
+    end
+  end
+end